Add files via upload
In this project, we designed a 3-bit binary multiplier circuit multiplied by a given 
constant 4. A binary multiplier is a special type of combinational logic circuit. It is 
used to multiply two binary numbers. In this process, the number that will be 
multiplied by the other number is referred to as the multiplicand, and the number 
that will be multiplied is referred to as the multiplier. The multiplicand is 
multiplied by each bit of the multiplier starting from the least significant bit. Each 
multiplication forms a partial product, successive partial products are shifted one 
position to the left. The final product is obtained from the sum of the partial 
products.